More libpi3hat timing improvements
Stop calling ::usleep, instead just busy wait as it is much more deterministic.

Add more options for minimum timing delays that results in us getting
all of the current CAN packets most of the time and not just stale
ones.

Stuff the auxiliary SPI with up to 3 bytes at a time.  It leaves dead
time between FIFO entries even when the FIFO is full, so minimizing
the number of FIFO entries used decreases the total time required.