Automatically recover from can Bus_Off state
The documentation on how to do this is basically non-existent, as is
anything in Google as of 2020-11-13.  Older versions of the ST CAN
peripheral had a mechanism to automatically recover from Bus_Off, but
no more.  It seems that just asking the peripheral to re-initialize by
clearing INIT in CCCR is sufficient to get it working, so we'll go
with that.  That is hinted at by the italicized note in '43.4.13 FDCAN
protocol status register (FDCAN_PSR)' of the reference manual.

While we are at it, expose some CAN error information over SPI.