Increase the setup and hold times
On aarch64, the kernel does something funny with the times returned
from clock_gettime.  I had to increase the times to 9us to avoid
sometimes having an actual delay of less than a microsecond.

The "official" way to do this is nanosleep, which has a minimum sleep
time of more like 20us.  I guess the next step would be to calibrate
some bogomips and use that.