Use synchronization barriers around busy waits
This does a better job of resolving timing irregularities on aarch64,
and is a more sound approach on armv7l as well.  Since our busy waits
are intended to induce delays between when, say the chip select line
changes and the SPI peripheral starts sending, re-orders across that
will negate their effect.  Use the strongest ARM synchronization
barrier on offer for either architecture to get the busy loops to
actually be isolated.

This lets us put the setup and hold times back down to 3us for all
platforms.

While we are at it, increase the DOUT hold time for the auxiliary SPI
bus, which improves its reliability significantly without much time
penalty.