Be more robust when no device acknowledges a frame
If the pi3hat attempts to send a CAN frame and no device acknowledges
it, the retranmissions continue indefinitely.  If the host keeps
enqueueing packets, eventually the hardware FIFO becomes full.  At
that point, the firmware would enter a busy loop preventing anything
further from working.

We resolve that in two ways.

First, if the hardware FIFO is full, just discard new frames sent from
the host.

Second, 50ms after the last frame was successfully enqueued with the
hardware, cancel all outstanding transmissions.  This ensures that a
single frame which is never acknowledged by anything does not prevent
later frames with an address that would be acknowledged from going
out.