Fix some cancellation safety issues
First, require that each _transaction not be cancelled.  Using the
pi3hat cpp interface, we don't expect any individual call to take very
long anyway as we have pretty tight timeouts.

Second, if a subscription handler is cancelled but a frame is received
anyway, stick it in the global queue so that it isn't lost.

Exception safety is hard.