Fixes to synchronization
asyncio.shield(coro()) has odd failure modes.  The documentation says
that internally it does an asyncio.create_task(coro()), but also that
you *must* keep a reference to the created task or it may be GC'd.
Even when I manually did the create_task and kept a reference, the
coroutine kept getting GC'd away releasing the lock.

Also acquire the lock before requesting the attitude.